Re-did room creation
Before, rooms would be created before the user entered in game
parameters. Now, rooms are created when the user clicks create, given
that all game parameters are entered.

Furthermore, 'small blind amount' game parameter was removed, replaced
with 'name'.

<Link> replaced with <Redirect> in order to allow a now async createRoom
function to work.

goldfire's pokersolver library added to dependencies, to be used in the
future.

Also added a todo.txt, so I can keep track of what needs to be done.